move OptimisticKnnVectorQuery into AbstractKnnVectorQuery
